pyproject.toml
[build-system]
requires = ["hatchling>=1.25"]
build-backend = "hatchling.build"

[project]
name = "lophos"
version = "0.1.0"
description = "LOops & Peaks HaplOtype phasing Suite — allele-specific phasing of CTCF peaks & loops from phased HiChIP BAMs."
readme = "README.md"
license = { file = "LICENSE" }
requires-python = ">=3.10"
authors = [{ name = "LOPHOS Devs" }]
keywords = ["HiChIP", "CTCF", "haplotype", "allele-specific", "3D genome", "loops", "peaks"]

dependencies = [
  "pysam>=0.22",
  "pandas>=2.2",
  "numpy>=1.26",
  "scipy>=1.11",
  "typer>=0.12",
  "rich>=13.7",
  "pyyaml>=6.0.1",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.0",
  "pytest-cov>=5.0",
  "ruff>=0.5",
  "black>=24.4",
  "mypy>=1.10",
  "pre-commit>=3.7",
]
motif = [
  "pyfaidx>=0.8",
  "biopython>=1.83"
]

[project.scripts]
lophos = "lophos.cli:app"

[tool.black]
line-length = 100
target-version = ["py310"]

[tool.ruff]
line-length = 100
target-version = "py310"
lint.select = ["E", "F", "I", "B", "UP", "ARG", "C90"]
lint.ignore = ["E501"]  # handled by black

[tool.mypy]
python_version = "3.10"
warn_unused_ignores = true
warn_redundant_casts = true
strict = true
exclude = ["tests/"]

[tool.hatch.build.targets.wheel]
packages = ["src/lophos"]
